# backend/openapi.yaml
openapi: 3.0.3

info:
  title: TDD Challenge API
  description: API for the TDD Challenge application.
  version: 1.0.0

servers:
  - url: http://localhost:4000
    description: Local development server

tags:
  - name: Health
    description: API health operations
  - name: Challenges
    description: TDD challenge operations

paths:

  /api/health:
    get:
      tags:
        - Health
      summary: Check API health
      description: Returns the current health status of the API.
      operationId: getHealth
      responses:
        "200":
          description: API is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
              example:
                status: ok

        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                message: Internal server error

  /api/challenges:
    get:
      tags:
        - Challenges
      summary: Get all challenges
      description: Returns all available TDD challenges.
      operationId: listChallenges
      responses:
        "200":
          description: List of available challenges
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Challenge"
              example:
                - id: react-counter
                  title: Build a Counter
                  description: Create a counter by starting with behavior tests, then implement the smallest solution.
                  difficulty: Beginner
                  concepts:
                    - React state
                    - Events
                    - Component tests
                  steps:
                    - Write the failing test
                    - Implement the counter
                    - Refactor

                - id: api-health
                  title: Health Endpoint
                  description: Build a predictable health endpoint from an API contract test.
                  difficulty: Intermediate
                  concepts:
                    - Express
                    - HTTP
                    - Supertest
                  steps:
                    - Write the HTTP test
                    - Implement GET /api/health
                    - Refactor

                - id: challenge-filter
                  title: Filter Challenges
                  description: Add filtering behavior with tests covering empty, partial, and exact matches.
                  difficulty: Advanced
                  concepts:
                    - TypeScript
                    - Pure functions
                    - Edge cases
                  steps:
                    - Define behavior with tests
                    - Implement filtering
                    - Refactor

        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                message: Internal server error

  /api/challenges/{id}:
    get:
      tags:
        - Challenges
      summary: Get challenge by ID
      description: Returns a single TDD challenge using its string ID.
      operationId: getChallenge
      parameters:
        - name: id
          in: path
          required: true
          description: Unique challenge ID.
          schema:
            type: string
          example: react-counter

      responses:
        "200":
          description: Challenge found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Challenge"
              example:
                id: react-counter
                title: Build a Counter
                description: Create a counter by starting with behavior tests, then implement the smallest solution.
                difficulty: Beginner
                concepts:
                  - React state
                  - Events
                  - Component tests
                steps:
                  - Write the failing test
                  - Implement the counter
                  - Refactor

        "404":
          description: Challenge not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                message: Challenge not found

        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                message: Internal server error

components:

  schemas:

    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: ok
      required:
        - status

    Challenge:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier of the challenge.
          example: react-counter

        title:
          type: string
          description: Challenge title.
          example: Build a Counter

        description:
          type: string
          description: Description of the challenge.
          example: Create a counter by starting with behavior tests, then implement the smallest solution.

        difficulty:
          type: string
          description: Difficulty level of the challenge.
          enum:
            - Beginner
            - Intermediate
            - Advanced
          example: Beginner

        concepts:
          type: array
          description: Concepts covered by the challenge.
          items:
            type: string
          example:
            - React state
            - Events
            - Component tests

        steps:
          type: array
          description: Steps involved in completing the challenge.
          items:
            type: string
          example:
            - Write the failing test
            - Implement the counter
            - Refactor

      required:
        - id
        - title
        - description
        - difficulty
        - concepts
        - steps

    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Error message.
          example: Challenge not found
      required:
        - message